Remove phantom recursive dependency in setup.py
This removes the "import phantom" at the top
of the "setup.py" file. This is important because when installing
phantom as a pip dependency, the "setup.py" file is executed
and python tries to import the module phantom, but it does
not exists yet because it's the module we are currently trying
to install with pip. This causes an error.

So the version number is moved to the setup.py and the
version for phantom is fetched in __init__.py from the
setup.py itself.